+++ reference/SampleCompany/NodeManagers/MemoryBuffer/MemoryBufferState.cs
@@ -176,7 +176,7 @@ namespace SampleCompany.NodeManagers.MemoryBuffer
         }
 
         /// <summary>
-        /// Creates an object which can browser the tags in the buffer.
+        /// Creates an object which can browse the tags in the buffer.
         /// </summary>
         public override INodeBrowser CreateBrowser(
             ISystemContext context,